requirements.txt
gradio>=4.0.0
numpy>=1.24.0
Pillow>=9.0.0
watchdog>=3.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0